+++ src/expression/compound_expression.ts
@@ -194,7 +194,7 @@ function rgba(ctx, [r, g, b, a]) {
 }
 
 function has(key, obj) {
-    return key in obj;
+    return key in obj && obj[key] !== undefined;
 }
 
 function get(key, obj) {
@@ -428,7 +428,15 @@ CompoundExpression.register(expressions, {
             return typeof a === typeof b && a >= b;
         }
     ],
-    'filter-has': [BooleanType, [ValueType], (ctx, [k]) => (k as any).value in ctx.properties()],
+    'filter-has': [
+        BooleanType,
+        [ValueType],
+        (ctx, [k]) => {
+            const key = (k as any).value;
+            const props = ctx.properties();
+            return key in props && props[key] !== undefined;
+        }
+    ],
     'filter-has-id': [BooleanType, [], (ctx) => ctx.id() !== null && ctx.id() !== undefined],
     'filter-type-in': [
         BooleanType,